requirements.txt
Flask==2.3.2
pymongo==4.6.1
dnspython==2.4.2
python-dotenv==1.0.1